refactor: rename component name
