+++ packages/settings/events/src/components/BindEventsDialogContent.vue
