+++ packages/settings/events/src/components/BindEventsDialogSidebar.vue
